cadastro habilidade

+++ src/main/java/dao/IUsuarioDao.java
@@ -7,5 +7,5 @@ import java.sql.SQLException;
 public interface IUsuarioDao {
 
     public void criarUsuario(String nome, String email, String senha, String tipo) throws SistemaRHDBException;
-    public long loginUsuario(String email,String senha) throws SistemaRHDBException;
+    public int loginUsuario(String email,String senha) throws SistemaRHDBException;
 }
